Extract bridge builders
Avoid using the ABI types almost everywhere
